% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{cleanCust}
\alias{cleanCust}
\title{Format customer names and addresses for fuzzy matching}
\usage{
cleanCust(x)
}
\arguments{
\item{x}{string}
}
\value{
String
}
\description{
Performs common transformations like making co, corp and corporation all
co; and converting all letters to lower case
}
\examples{
cleanCust('3M Corp., Suite #18, Innovation Boulevard')
}
