% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{dollarChar}
\alias{dollarChar}
\title{dollarChar}
\usage{
dollarChar(x = "dollars", ndecimal = 2, unit = "d", currency = "$",
  direction = FALSE)
}
\arguments{
\item{x}{Numeric. Dollar amount to be formatted.}

\item{ndecimal}{Integer. Number of decimal places to round to}

\item{unit}{What units to convert x to
\itemize{
  \item d = dollars (ones)
  \item k = thousands
  \item m = millions
  \item b = billions
  \item t = trillions (this is a lot of money, are you sure you
  don't have a data issue?)
}}

\item{currency}{String. Symbol to use as prefix of number.}

\item{direction}{Boolean. Should + proceed the number for
positive numbers?}
}
\value{
String.
}
\description{
dollarChar
}
\examples{
dollarChar(x = 2118999, unit = 'm')
}
