% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_index.R
\name{make_month_index}
\alias{make_month_index}
\title{Create a month index for analysis and plotting}
\usage{
make_month_index(df, SLS_YR, SLS_MNTH_INT)
}
\arguments{
\item{df}{Data Frame: Data set you want an index for}

\item{SLS_YR}{Bare Column Name: Name of column containing the year as integer}

\item{SLS_MNTH_INT}{Bare Column Name: Name of column containing the month as integer}
}
\value{
Data Frame containing the columns provided as SLS_YR, SLS_MNTH_INT and a index
}
\description{
Create a month index for analysis and plotting
}
\examples{
make_month_index(data.frame(SLS_YR = c(2016, 2017),
SLS_MNTH_INT = c(10,4)),
SLS_YR = SLS_YR, SLS_MNTH_INT = SLS_MNTH_INT)
}
