% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscFormatR.R
\docType{package}
\name{miscFormatR}
\alias{miscFormatR}
\alias{miscFormatR-package}
\title{miscFormatR: Useful formatting functions I couldn't find in other packages}
\description{
The miscFormatR package provides functions to automate tedious formatting tasks
}
\section{Save time, be consistent}{

\itemize{
  \item \code{\link{dollarChar}}
  \item \code{\link{copywriteDate}}
  \item \code{\link{cleanCust}}
  \item \code{\link{make_month_index}}
}
}

