% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{percentChar}
\alias{percentChar}
\title{Convert percents to string representations}
\usage{
percentChar(percent, ndecimal = 1)
}
\arguments{
\item{percent}{a percent, like 0.63}

\item{ndecimal}{number of decimal places to round  to}
}
\value{
string
}
\description{
Convert percents to string representations
}
\examples{
percentChar(.05364, ncecimal = 2)
}
